requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.31
bcrypt==4.1.3
blinker==1.8.2
Flask-SQLAlchemy==3.1.1
psycopg==3.1.19
psycopg-binary==3.1.19
python-dotenv==1.0.1
Jinja2==3.1.4
